"Few" changes ...

+++ app/Http/Controllers/ProjectController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use App\Models\Project;
+
 class ProjectController extends Controller
 {
     public function index()
@@ -78,6 +79,8 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
+        abort_if($project->owner_id !== auth()->id(), 403);
+
         $schema = (new Project())->schema;
         $item = $project;
 
@@ -86,6 +89,8 @@ class ProjectController extends Controller
 
     public function update(Request $request, Project $project)
     {
+        abort_if($project->owner_id !== auth()->id(), 403);
+
         $schema = (new Project())->schema;
         $validated = $this->validateSchema($request, $schema);
 
@@ -107,6 +112,8 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        abort_if($project->owner_id !== auth()->id(), 403);
+
         $schema = (new Project())->schema;
         $project->delete();
